requirements.txt
asgiref
async-timeout
autopep8
beautifulsoup4
certifi
cffi
chardet
charset-normalizer
cloudinary
crispy-bootstrap5
cryptography
cssselect2
defusedxml
Deprecated
diff-match-patch
distlib
dj-database-url
Django
django-allauth
django-ckeditor
django-cloudinary-storage
django-crispy-forms
django-filer
django-import-export
django-js-asset
django-mptt
django-multiurl
django-polymorphic
django-ranged-response
django-redis
django-rename-app
django-simple-captcha
django-tinymce
easy-thumbnails
et-xmlfile
filelock
flake8
gunicorn
idna
lxml
MarkupPy
mccabe
oauthlib
odfpy
openpyxl
packaging
pillow
platformdirs
psycopg2
psycopg2-binary
pycodestyle
pycparser
pyflakes
PyJWT
pyparsing
python3-openid
pytz
PyYAML
redis
reportlab
requests
requests-oauthlib
six
soupsieve
sqlparse
svglib
tablib
tinycss2
toml
typing_extensions
Unidecode
urllib3
virtualenv
webencodings
whitenoise
wrapt
xlrd
xlwt